+++ example/lib/main.dart
@@ -3,7 +3,6 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:install_plugin/install_plugin.dart';
-import 'package:permission_handler/permission_handler.dart';
 
 void main() => runApp(new MyApp());
 
@@ -14,6 +13,7 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   String _platformVersion = 'Unknown';
+  String _appUrl = '';
 
   @override
   void initState() {
@@ -21,19 +21,13 @@ class _MyAppState extends State<MyApp> {
     initPlatformState();
   }
 
-  // Platform messages are asynchronous, so we initialize in an async method.
   Future<void> initPlatformState() async {
     String platformVersion;
-    // Platform messages may fail, so we use a try/catch PlatformException.
     try {
       platformVersion = await InstallPlugin.platformVersion;
     } on PlatformException {
       platformVersion = 'Failed to get platform version.';
     }
-
-    // If the widget was removed from the tree while the asynchronous platform
-    // message was in flight, we want to discard the reply rather than calling
-    // setState to update our non-existent appearance.
     if (!mounted) return;
 
     setState(() {
@@ -50,14 +44,20 @@ class _MyAppState extends State<MyApp> {
         ),
         body: new Column(
           children: <Widget>[
+            Text(_platformVersion),
             FlatButton(
                 onPressed: () {
                   onClickInstallApk();
                 },
                 child: Text('install')),
-            FlatButton(onPressed: (){
-              onClickGotoAppStore();
-            }, child: Text('gotoAppStore'))
+            TextField(
+              decoration:
+                  InputDecoration(hintText: 'URL for app store to launch'),
+              onChanged: (url) => _appUrl = url,
+            ),
+            FlatButton(
+                onPressed: () => onClickGotoAppStore(_appUrl),
+                child: Text('gotoAppStore'))
           ],
         ),
       ),
@@ -65,30 +65,19 @@ class _MyAppState extends State<MyApp> {
   }
 
   void onClickInstallApk() {
-    requestPermission(PermissionGroup.storage).then((result) {
-      print('requestPermission result: $result');
-      if (result[PermissionGroup.storage] == PermissionStatus.granted) {
-        InstallPlugin.installApk('/storage/emulated/0/zpartner/update.apk',
-                'com.zaihui.installpluginexample')
-            .then((result) {
-          print('install apk $result');
-        }).catchError((error) {
-          print('install apk error: $error');
-        });
-      } else {}
+    InstallPlugin.installApk('/storage/emulated/0/zpartner/update.apk',
+            'com.zaihui.installpluginexample')
+        .then((result) {
+      print('install apk $result');
     }).catchError((error) {
-      print('requestPermission error: $error');
+      print('install apk error: $error');
     });
   }
 
-  void onClickGotoAppStore() {
-    InstallPlugin.gotoAppStore('https://itunes.apple.com/cn/app/%E5%86%8D%E6%83%A0%E5%90%88%E4%BC%99%E4%BA%BA/id1375433239?l=zh&ls=1&mt=8');
+  void onClickGotoAppStore(String url) {
+    url = url.isEmpty
+        ? 'https://itunes.apple.com/cn/app/%E5%86%8D%E6%83%A0%E5%90%88%E4%BC%99%E4%BA%BA/id1375433239?l=zh&ls=1&mt=8'
+        : url;
+    InstallPlugin.gotoAppStore(url);
   }
-  
-  Future<Map<PermissionGroup, PermissionStatus>> requestPermission(
-      PermissionGroup permission) async {
-    final List<PermissionGroup> permissions = <PermissionGroup>[permission];
-    return await PermissionHandler().requestPermissions(permissions);
-  }
-
 }