+++ CHANGELOG.md
@@ -1,3 +1,15 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Add `installApk` function for android to install apk, supporting ANDROID.O.
+* Add `gotoAppStore` function for iOS to launch a url to App Store.
+
+## 0.0.2
+
+* Fix the bug that Pixel Device install apk fail.
+* Optimize the user experience when app asking for `Settings.ACTION_MANAGE_UNKNOWN_APP_SOURCES` permission for Android device above ANDROID.O.
+
+## 1.0.0
+
+* Add CHANGELOG and README.
+* Remove unused code.
+* Update plugin homepage url.